# ixor optimisation