Speed up ixor between bitmaps