@richardstartin 2021-11-26
Here are some fairly noisy benchmark results though this change is motivated by the code which gets generated more than anything else.

before
```
Benchmark                               (thatDensity)  (thisDensity)  Mode  Cnt    Score     Error  Units
BitmapContainerXorBenchmark.stable                0.1            0.1  avgt    5  887.426 ±  31.657  ns/op
BitmapContainerXorBenchmark.stable                0.1            0.5  avgt    5  940.839 ± 183.813  ns/op
BitmapContainerXorBenchmark.stable                0.5            0.1  avgt    5  900.977 ±  34.971  ns/op
BitmapContainerXorBenchmark.stable                0.5            0.5  avgt    5  903.406 ±  14.553  ns/op
BitmapContainerXorBenchmark.tendToZero            0.1            0.1  avgt    5  419.446 ±  10.448  ns/op
BitmapContainerXorBenchmark.tendToZero            0.1            0.5  avgt    5  420.538 ±  16.545  ns/op
BitmapContainerXorBenchmark.tendToZero            0.5            0.1  avgt    5  417.096 ±   1.289  ns/op
BitmapContainerXorBenchmark.tendToZero            0.5            0.5  avgt    5  417.373 ±   3.359  ns/op
```

after
```
Benchmark                               (thatDensity)  (thisDensity)  Mode  Cnt    Score    Error  Units
BitmapContainerXorBenchmark.stable                0.1            0.1  avgt    5  851.225 ±  7.401  ns/op
BitmapContainerXorBenchmark.stable                0.1            0.5  avgt    5  856.114 ± 11.096  ns/op
BitmapContainerXorBenchmark.stable                0.5            0.1  avgt    5  855.182 ± 19.634  ns/op
BitmapContainerXorBenchmark.stable                0.5            0.5  avgt    5  859.893 ± 11.731  ns/op
BitmapContainerXorBenchmark.tendToZero            0.1            0.1  avgt    5  327.902 ±  6.521  ns/op
BitmapContainerXorBenchmark.tendToZero            0.1            0.5  avgt    5  327.523 ±  6.597  ns/op
BitmapContainerXorBenchmark.tendToZero            0.5            0.1  avgt    5  326.116 ±  2.613  ns/op
BitmapContainerXorBenchmark.tendToZero            0.5            0.5  avgt    5  325.384 ±  0.353  ns/op
```