@lemire 2021-11-26
It is obviously good. :-)