@richardstartin 2021-11-26
For other in place operations, if the bitmap is in lazy mode the cardinality calculation is skipped and the `this` is returned. It might be worth doing that here too.